skip secret length check in OTP client setting (#5)
Some providers don't follow the RFCs to the letter and only use 10 bytes for the shared secret. Therefore, the generate() method now skips the secret length validation.

The verify() method continues to throw if the secret is too short. This is to enforce the security recommendations from the RFCs when you're implementing an OTP provider.